Change how we run the tests in ant.
There's now a single 'test' target, which omits groups of tests depending on the
value of the 'dont.test' system property, so to only run the maven tests you
could for example call
  ant test -Ddont.test=ast,exec,doc

Currently dont.test can include the words 'ast', 'exec', 'doc' and 'maven'.

The dependencies are also set up so that you don't spend time building the
tests .car if you're not going to run the 'exec' tests.