annotations: AbstractModelLoader: Fix a bootstrap problem.
Generating module and packages descriptors with annotations from the
language module revealed a problem when doing "ant clean publish"
in ceylon.language (but not when doing an "ant siblings package" in ceylonn-dist).

At the start of compilation the model loader tries to load the package descriptors.
When compiling ceylon.language it seemed to be loading the model from .class, which
 triggered loading of Shared$annotation and we fell over with
"Trying to load new source file after CeylonEnter has been called".