Issue #1170: Change error handling in CeylonCompileTool and Main.
Revise CompilerErrorException, CompilerBugException and SystemErrorException so they don't use @NonFatal

CeylonEnter & CeylonLog: Account for ceylon codegen errors separately from analysis errors, so we can properly distinguish between analysis errors, bugs in our codegen and bugs in later javac phases. We need to do this to properly distinguish typechecker error cases from compiler bug cases. Do this by using and keeping a count of different message keys in CeylonLog.

Main: Use an ExitState class to save adding lots of fields to Main and to centralize the logic for figuring how when something is an comtpile time error or a bug.