Implementation of `hash` drops bits (#1334)
Made hash into a long everywhere except where it is returned from hashCode(), this way internally we
stay in Ceylon land until we return it from hashCode for Java interop. This means that hash fields
are not stored as longs and hash initialiser parameters are also now longs, which I think is more
logical.