Fix #1935.
Because this affects types used in declarations this change breaks BC.

Instead of going raw when ALL the type arguments are intersections or unions do so when ANY are.
But simplify the type first
And when we've gone raw in a parameter then when we transform a method that
refines we need to also go raw in the parameter too.