Fix for #2033: Incompatible types error when extending MutableList
Generate extra abstract methods in interface which satisfy two interfaces with different type
arguments, if the methods (or attributes or member classes) have different return types. This
uses the same machinery as the mixin bridge generator, but with abstract methods, so that the
signatures are correct. This lets javac stop checking those interfaces.