CMRTests: fix test for #1104
Now, I'll admit I don't understand what changed to make this work, but it's rather logical
that it works and I don't understand how it could not have worked before:

- camel-jetty imports camel-http but does not share it, so:
- when we import camel-jetty we don't add camel-http to the classpath
- we import camel-http manually, which adds it to the classpath
- a type in camel-jetty has a supertype in camel-http
- when camel-jetty requires loading that supertype we find it in the classpath, and
- we check if it is imported and it is (even though not shared)

So it just all works. It's rather counter-intuitive because we don't actually
use anything from camel-http but the compiler is actually a flat classpath even
though it respects dependency imports and does not care about shared imports.

It would not have worked if camel-jetty got camel-http via a middle transitive dep though.